Restructure identity to use maps and lists instead of objects
It will finally compile and run again! In preparation for making ponies
from JSON files, I've changed the way that manually created ponies are
stored to match the way that data will be read in from the .json files.
This was something I had already started doing in February and May
before taking a long break from the project, but I never finished it.
Now things are coming along quite nicely.

I think starting to use git for this project was a good idea.