Change the priority queue to pop off the entry with the most stamina
I accidentally implemented the priority queue incorrectly. The smallest
value was being popped from the queue, which actually resulted in extra
retreads of the same nodes. There aren't a lot of characters in the test
room, however, so it didn't result in any significant performance drops.
Now, the most efficient search path is properly popped off of the queue
first.