Don't requeue nodes already in queue + readability changes
When a more efficient path to a tile is found before it's ever popped
off of the queue, it shouldn't be added to the queue a second time.
Check if a node is already in the queue before adding it and, if it
is, change its priority to reflect the stamina cost of the new
optimal path to that node.

The queuing of nodes can be collapsed into the update_move_node()
script, because there is no instance where a move node will be
created or updated without needing to explore or re-explore its
movement options around itself.

Improve some commenting in the move map creation code.

Clean up the organization of the files (the change to the overall
project file), order them to according to when they're called in the
recursive loop.

Lastly, remove 2 scripts that are no longer necessary, as they were
written for an older version of the priority queue that was scrapped
a long time ago.