Make path directions match elsewhere in code, plus other small tweaks
Paths used to use different numbers for representing directions than the
move maps. It was a very shortsighted decision, where I had made them
correspond directly to the index of the correct arrow sprite for that
direction, and the inconsistency in representation made debugging
confusing. Finally, I have amended my mistake, and brought some additional
changes along as well.

For instance, I improved the descriptiveness of the comment that is meant
to explain how the path segment graphics are determined in
move_path_tile_update(), while also moving the comments to be in-line
with the code, rather than in a massive block before it.

I also reorganized move_path_tile_get_dir() to not have a duplicated code
fragment.

Thanks to the consistent direction representation, I could change
most_efficient_path() to simply get the opposite direction of the move
node's optimal direction, instead of manually assigning each one.

Additionally, opposite_direction() only needs to multiply the direction
by -1. Cases are not necessary.