Use a global game object and game states to simplify game flow
This change adds a singleton game object as a global variable. The game
state property, which can be accessed with global.game.state, can dictate
the way that characters, cursors, menus, etc. should behave. Before this
change, all determinations on game state were made by checking certain
flags and properties of the cursor object. As such, it wasn't very
readable, and it was quite complicated and confusing to implement new
features and game states. By centralizing game states within a master
object, it will become much easier to read and manage the flow of the
program, and I won't have to define an unmanageable mess of flags
throughout different objects.

The changes include:
- Adding the game object, which stores game states and turn data. The turn
  data is currently unused.
- Add the game object to the test room, and move all the spriteless
  utility objects to outside the room.
- Add a game_set_state() function.
- Make clicking buttons and clicking objects with the cursor update the
  game state correctly
- Make cursor behavior check the game state to determine behavior.
- Check for the "animating" game state when determining if an obj_Pony
  instance should move.
- Remove all references to the isPaused property of the cursor, as now the
  game state is used to determine whether the in-game cursor should follow
  the mouse or respond to mouse clicks.
- Change the name of the "Cancel" button to "Close" to reflect its
  purpose of closing the menu.
- Change the size of the menu buttons to match the size of the custom font
  that I'm working on.
- Improve comments on cursor code and add more of them.