[gameplay] Basic submenu behavior and act mechanics
Clicking a character will now pull up a menu. This menu is currently
fixed in the top-left corner of the screen, but will later appear
directly next to the clicked character.

From the menu, you can choose to move the character, or to cancel the
selection. While a submenu is open, the in-game cursor will not follow
the mouse cursor. Choosing "Move" will activate the move map as
simply clicking the character once had. Choosing "Cancel" effectively
does nothing. No matter which option is clicked, the submenu is closed

Pressing the "E" key on the keyboard while highlighting a character
will bring up a temporary test mode of the attack system, with that
character designated as the "user". It automatically selects the
"user"'s first action and activates that action's sight map.
Pressing "E" again while highlighting a different character will
perform the action.

Added a distinction between Line of Sight and Same Elevation.
Line of Sight attacks like Magic Missile should be able to peek one
tile over into tiles of higher elevation. Same Elevation attacks like
Buck should require the characters to be on even ground.